backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
cryptography==41.0.7
pqcrypto==0.3.4
requests==2.31.0
python-multipart==0.0.6
passlib==1.7.4
bcrypt==4.1.2
python-jose==3.3.0
aiosqlite==0.19.0
secure-smtplib==0.1.1
authlib==1.6.4
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-api-python-client==2.182.0